refactor: add packetGroup controller/service
