+++ api/app/src/main/kotlin/packit/controllers/PacketController.kt
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.*
 import packit.model.PacketMetadata
 import packit.model.PageablePayload
 import packit.model.dto.PacketDto
-import packit.model.dto.PacketGroupDto
 import packit.model.dto.PacketGroupSummary
 import packit.model.toDto
 import packit.service.PacketService
@@ -75,16 +74,5 @@ class PacketController(private val packetService: PacketService)
             .ok()
             .headers(response.second)
             .body(response.first)
-    }
-
-    @GetMapping("/packetGroup")
-    fun getPacketGroups(
-        @RequestParam(required = false, defaultValue = "0") pageNumber: Int,
-        @RequestParam(required = false, defaultValue = "50") pageSize: Int,
-        @RequestParam(required = false, defaultValue = "") filterName: String,
-    ): ResponseEntity<Page<PacketGroupDto>>
-    {
-        val payload = PageablePayload(pageNumber, pageSize)
-        return ResponseEntity.ok(packetService.getPacketGroups(payload, filterName).map { it.toDto() })
     }
 }
