+++ api/app/src/main/kotlin/packit/service/PacketGroupService.kt
@@ -1,0 +1,26 @@
+package packit.service
+
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Sort
+import org.springframework.stereotype.Service
+import packit.helpers.PagingHelper
+import packit.model.PacketGroup
+import packit.model.PageablePayload
+import packit.repository.PacketGroupRepository
+
+interface PacketGroupService
+{
+    fun getPacketGroups(pageablePayload: PageablePayload, filteredName: String): Page<PacketGroup>
+}
+
+@Service
+class BasePacketGroupService(
+    private val packetGroupRepository: PacketGroupRepository
+) : PacketGroupService
+{
+    override fun getPacketGroups(pageablePayload: PageablePayload, filteredName: String): Page<PacketGroup>
+    {
+        val packetGroups = packetGroupRepository.findAllByNameContaining(filteredName, Sort.by("name"))
+        return PagingHelper.convertListToPage(packetGroups, pageablePayload)
+    }
+}
