+++ api/app/src/test/kotlin/packit/integration/controllers/PacketControllerTest.kt
@@ -1,6 +1,5 @@
 package packit.integration.controllers
 
-import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
@@ -14,10 +13,6 @@ import org.springframework.http.ResponseEntity
 import packit.integration.IntegrationTest
 import packit.integration.WithAuthenticatedUser
 import packit.model.Packet
-import packit.model.PacketGroup
-import packit.model.dto.PacketGroupDto
-import packit.model.toDto
-import packit.repository.PacketGroupRepository
 import packit.repository.PacketRepository
 import kotlin.test.assertEquals
 
@@ -25,13 +20,9 @@ import kotlin.test.assertEquals
 class PacketControllerTest : IntegrationTest()
 {
     @Autowired
-    private lateinit var packetGroupRepository: PacketGroupRepository
-
-    @Autowired
     private lateinit var packetRepository: PacketRepository
-    private lateinit var packetGroups: List<PacketGroup>
     private lateinit var packets: List<Packet>
-    private val packetNames = listOf(
+    private val packetGroupNames = listOf(
         "test-packetGroupName-1",
         "test-packetGroupName-2",
         "test-packetGroupName-3",
@@ -42,20 +33,11 @@ class PacketControllerTest : IntegrationTest()
     @BeforeAll
     fun setupData()
     {
-        packetGroups = packetGroupRepository.saveAll(
-            listOf(
-                PacketGroup(name = packetNames[1]),
-                PacketGroup(name = packetNames[0]),
-                PacketGroup(name = packetNames[4]),
-                PacketGroup(name = packetNames[3]),
-                PacketGroup(name = packetNames[2]),
-            )
-        )
         packets = packetRepository.saveAll(
             listOf(
                 Packet(
                     "20230427-150755-2dbede93",
-                    packetNames[0],
+                    packetGroupNames[0],
                     "test-packetGroupName-1",
                     mapOf("name" to "value"),
                     false,
@@ -65,8 +47,8 @@ class PacketControllerTest : IntegrationTest()
                 ),
                 Packet(
                     "20230427-150755-2dbede94",
-                    packetNames[0],
-                    packetNames[0],
+                    packetGroupNames[0],
+                    packetGroupNames[0],
                     mapOf("a" to 1),
                     false,
                     0.0,
@@ -75,7 +57,7 @@ class PacketControllerTest : IntegrationTest()
                 ),
                 Packet(
                     "20230427-150755-2dbede95",
-                    packetNames[2],
+                    packetGroupNames[2],
                     "test-packetGroupName-3",
                     mapOf("alpha" to true),
                     false,
@@ -85,7 +67,7 @@ class PacketControllerTest : IntegrationTest()
                 ),
                 Packet(
                     "20230427-150755-2dbede96",
-                    packetNames[3],
+                    packetGroupNames[3],
                     "test-packetGroupName-4",
                     mapOf(),
                     true,
@@ -95,7 +77,7 @@ class PacketControllerTest : IntegrationTest()
                 ),
                 Packet(
                     "20230427-150755-2dbede97",
-                    packetNames[4],
+                    packetGroupNames[4],
                     "test-packetGroupName-5",
                     mapOf(),
                     true,
@@ -110,7 +92,6 @@ class PacketControllerTest : IntegrationTest()
     @AfterAll
     fun cleanup()
     {
-        packetGroupRepository.deleteAll(packetGroups)
         packetRepository.deleteAll(packets)
     }
 
@@ -215,48 +196,7 @@ class PacketControllerTest : IntegrationTest()
 
         assertHtmlFileSuccess(result)
     }
-
-    @Test
-    @WithAuthenticatedUser(authorities = ["packet.read"])
-    fun `get ordered pageable packetGroups with filtered name`()
-    {
-        val result: ResponseEntity<String> = restTemplate.exchange(
-            "/packets/packetGroup?pageNumber=0&pageSize=10&filterName=test-packetGroupName",
-            HttpMethod.GET,
-            getTokenizedHttpEntity()
-        )
-        assertSuccess(result)
-        val resultPacketGroups = jacksonObjectMapper().readTree(result.body).get("content")
-            .let {
-                jacksonObjectMapper().convertValue(
-                    it,
-                    object : TypeReference<List<PacketGroupDto>>()
-                    {}
-                )
-            }
-        assert(resultPacketGroups.containsAll(packetGroups.map { it.toDto() }))
-        assertEquals(5, resultPacketGroups.size)
-        assertEquals("test-packetGroupName-1", resultPacketGroups[0].name)
-        assertEquals("test-packetGroupName-5", resultPacketGroups[resultPacketGroups.size - 1].name)
-    }
-
-    @Test
-    @WithAuthenticatedUser(authorities = ["packet.read"])
-    fun `return correct page information for get pageable packet groups `()
-    {
-        val result: ResponseEntity<String> = restTemplate.exchange(
-            "/packets/packetGroup?pageNumber=0&pageSize=10&filterName=test-packetGroupName",
-            HttpMethod.GET,
-            getTokenizedHttpEntity()
-        )
-        assertSuccess(result)
-        val resultPage = jacksonObjectMapper().readTree(result.body)
-        assertEquals(5, resultPage.get("totalElements").asInt())
-        assertEquals(1, resultPage.get("totalPages").asInt())
-        assertEquals(0, resultPage.get("number").asInt())
-        assertEquals(10, resultPage.get("size").asInt())
-    }
-
+    
     @Test
     @WithAuthenticatedUser(authorities = ["packet.read:packet:test-packetGroupName-1:20230427-150755-2dbede93"])
     fun `findPacketMetadata returns metadata if user has correct specific permission`()
@@ -389,36 +329,6 @@ class PacketControllerTest : IntegrationTest()
     {
         val result: ResponseEntity<String> = restTemplate.exchange(
             "/packets/packetGroupSummary",
-            HttpMethod.GET,
-            getTokenizedHttpEntity()
-        )
-        assertEquals(2, jacksonObjectMapper().readTree(result.body).get("totalElements").asInt())
-    }
-
-    @Test
-    @WithAuthenticatedUser(authorities = ["packet.read:packetGroup:random-name"])
-    fun `getPacketGroups returns empty page if no permissions match`()
-    {
-        val result: ResponseEntity<String> = restTemplate.exchange(
-            "/packets/packetGroup",
-            HttpMethod.GET,
-            getTokenizedHttpEntity()
-        )
-
-        assertEquals(0, jacksonObjectMapper().readTree(result.body).get("totalElements").asInt())
-    }
-
-    @Test
-    @WithAuthenticatedUser(
-        authorities = [
-            "packet.read:packetGroup:test-packetGroupName-1",
-            "packet.read:packet:test-packetGroupName-3:20230427-150755-2dbede95"
-        ]
-    )
-    fun `getPacketGroups returns of packet groups user can see`()
-    {
-        val result: ResponseEntity<String> = restTemplate.exchange(
-            "/packets/packetGroup",
             HttpMethod.GET,
             getTokenizedHttpEntity()
         )
