+++ api/app/src/test/kotlin/packit/unit/service/PacketGroupServiceTest.kt
@@ -1,0 +1,36 @@
+package packit.unit.service
+
+import org.junit.jupiter.api.Test
+import org.mockito.kotlin.*
+import org.springframework.data.domain.Sort
+import packit.model.PacketGroup
+import packit.model.PageablePayload
+import packit.repository.PacketGroupRepository
+import packit.service.BasePacketGroupService
+import kotlin.test.assertEquals
+
+class PacketGroupServiceTest
+{
+    @Test
+    fun `getPacketGroups calls repository with correct params and returns its result`()
+    {
+        val packetGroupRepository = mock<PacketGroupRepository>()
+        val sut = BasePacketGroupService(packetGroupRepository)
+        val pageablePayload = PageablePayload(0, 10)
+        val filterName = "test"
+        val packetGroups = listOf(PacketGroup("test1"), PacketGroup("test2"))
+        whenever(
+            packetGroupRepository.findAllByNameContaining(
+                eq(filterName), any<Sort>()
+            )
+        ).thenReturn(packetGroups)
+
+        val result = sut.getPacketGroups(pageablePayload, filterName)
+
+        assertEquals(packetGroups, result.content)
+        verify(packetGroupRepository).findAllByNameContaining(
+            filterName,
+            Sort.by("name")
+        )
+    }
+}
