+++ app/src/app/components/contents/manageAccess/updatePermission/hooks/useGetResourcesForScopedPermissions.ts
@@ -4,10 +4,17 @@ import { PermissionScope } from "../../../../../../lib/constants";
 import { PageableBasicDto } from "../../../../../../types";
 import { fetcher } from "../../../../../../lib/fetch";
 
+const scopeToPathVariable = {
+  global: null,
+  packet: "packets",
+  tag: "tag",
+  packetGroup: "packetGroup"
+};
+
 export const useGetResourcesForScopedPermissions = (scope: PermissionScope, filterName: string) => {
-  const scopePathVariable = scope === "tag" ? "tag" : `packets${scope === "packetGroup" ? "/packetGroup" : ""}`;
+  const scopePathVariable = scopeToPathVariable[scope];
   const { data, isLoading, error } = useSWR<PageableBasicDto>(
-    scope !== "global" ? `${appConfig.apiUrl()}/${scopePathVariable}?filterName=${filterName}&pageSize=150` : null,
+    scopePathVariable ? `${appConfig.apiUrl()}/${scopePathVariable}?filterName=${filterName}&pageSize=150` : null,
     (url: string) => fetcher({ url })
   );
