+++ app/src/msw/handlers/packetHandler.ts
@@ -1,11 +1,17 @@
 import { rest } from "msw";
 import appConfig from "../../config/appConfig";
-import { mockPacket } from "../../tests/mocks";
+import { mockPacket, mockPacketGroupResponse } from "../../tests/mocks";
 
-const specificPacketIndexUri = `${appConfig.apiUrl()}/packets/metadata/${mockPacket.id}`;
+const packetIndexUri = `${appConfig.apiUrl()}/packets`;
 
 export const packetHandlers = [
-  rest.get(specificPacketIndexUri, (req, res, ctx) => {
+  rest.get(`${packetIndexUri}/metadata/${mockPacket.id}`, (req, res, ctx) => {
     return res(ctx.json(mockPacket));
+  }),
+  rest.get(`${packetIndexUri}/${mockPacketGroupResponse.content[0].name}`, (req, res, ctx) => {
+    return res(ctx.json(mockPacketGroupResponse));
+  }),
+  rest.get(`${packetIndexUri}`, (req, res, ctx) => {
+    return res(ctx.json(mockPacketGroupResponse));
   })
 ];
